refactor: order-api 모듈 영역 적용
