--- src/main/java/kitchenpos/tablegroup/domain/OrderTableEmpty.java
@@ -1,33 +1,0 @@
-package kitchenpos.tablegroup.domain;
-
-import kitchenpos.common.ErrorCode;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class OrderTableEmpty {
-    @Column
-    private boolean empty;
-
-    protected OrderTableEmpty() {}
-
-    public OrderTableEmpty(boolean empty) {
-        this.empty = empty;
-    }
-
-    public void validateForTableGrouping() {
-        if (!this.empty) {
-            throw new IllegalArgumentException(
-                    ErrorCode.NON_EMPTY_ORDER_TABLE_CANNOT_BE_INCLUDED_IN_TABLE_GROUP.getErrorMessage());
-        }
-    }
-
-    public OrderTableEmpty changeEmpty(boolean empty) {
-        return new OrderTableEmpty(empty);
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-}
